Code for Random Letter Generation
This code can randomly generate an array of letters without repetitions. This code takes in a integer as an argument. Said integer is the number of letters required. Code returns array of letters. Usually, a vowel is included in the randomized array of letters, but another version which makes sure there are at least two vowels are included will be uploaded soon.